group replay utilities
