+++ src/gem/replays/__init__.py
@@ -1,0 +1,19 @@
+"""Replay utility helpers for fetching and batch-processing replay files."""
+
+from gem.replays.batch import (
+    ParseResult,
+    parse_many,
+    parse_many_to_dataframe,
+    parse_many_to_parquet,
+)
+from gem.replays.fetch import download_and_decompress, fetch_replay, fetch_replay_url
+
+__all__ = [
+    "ParseResult",
+    "download_and_decompress",
+    "fetch_replay",
+    "fetch_replay_url",
+    "parse_many",
+    "parse_many_to_dataframe",
+    "parse_many_to_parquet",
+]
